tools/discord_mcp/requirements.txt
discord.py>=2.3.0
mcp>=1.2.1